requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.9.0=pypi_0
alabaster=0.7.12=pypi_0
arrow=0.15.6=pypi_0
astor=0.8.1=pypi_0
astroid=2.4.1=py37_0
astunparse=1.6.3=pypi_0
attrs=19.3.0=pypi_0
babel=2.8.0=pypi_0
backcall=0.1.0=pypi_0
blas=1.0=mkl
bleach=3.1.5=pypi_0
ca-certificates=2020.1.1=0
cachetools=4.1.0=pypi_0
certifi=2020.4.5.1=py37_0
chardet=3.0.4=pypi_0
cloudpickle=1.4.1=py_0
colorama=0.4.3=py_0
cycler=0.10.0=pypi_0
cython=0.29.19=pypi_0
cytoolz=0.10.1=py37he774522_0
dask-core=2.17.2=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=pypi_0
docutils=0.16=pypi_0
entrypoints=0.3=pypi_0
freetype=2.9.1=ha9979f8_1
gast=0.2.2=pypi_0
google-auth=1.16.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.29.0=pypi_0
h5py=2.10.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=vc14hc45fdbb_0
idna=2.9=pypi_0
imageio=2.8.0=py_0
imagesize=1.2.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=1.6.0=pypi_0
imutils=0.5.3=pypi_0
intel-openmp=2020.1=216
ipykernel=5.3.0=pypi_0
ipyparallel=6.3.0=pypi_0
ipython=7.15.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
isort=4.3.21=py37_0
jedi=0.17.0=pypi_0
jinja2=2.11.2=pypi_0
jpeg=9b=vc14h4d7706e_1
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.3=pypi_0
jupyter-core=4.6.3=pypi_0
keras=2.2.4=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=py37h74a9793_0
lazy-object-proxy=1.4.3=py37he774522_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_0
markdown=3.2.2=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.2.1=pypi_0
matplotlib-base=3.1.3=py37h64f37c6_0
mccabe=0.6.1=py37_1
mistune=0.8.4=pypi_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.6=pypi_0
networkx=2.4=py_0
nose=1.3.7=pypi_0
notebook=6.0.3=pypi_0
numpy=1.18.5=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1g=he774522_0
opt-einsum=3.2.1=pypi_0
packaging=20.4=pypi_0
pandas=1.0.4=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.1.2=py37hcc1f983_0
pip=20.0.2=py37_3
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.5=pypi_0
protobuf=3.12.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.6.1=pypi_0
pylint=2.5.2=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37ha878b3d_0
pyrsistent=0.16.0=pypi_0
python=3.7.7=h81c818b_4
python-dateutil=2.8.1=py_0
pytz=2020.1=pypi_0
pywavelets=1.1.1=py37he774522_0
pywin32=227=pypi_0
pywinpty=0.5.7=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.1=pypi_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.7.4=pypi_0
qtpy=1.9.0=pypi_0
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scikit-image=0.14.2=pypi_0
scipy=1.4.1=py37h9439919_0
send2trash=1.5.0=pypi_0
setuptools=46.4.0=py37_0
shapely=1.7.0=pypi_0
sip=4.19.13=py37ha925a31_0
six=1.15.0=pypi_0
snowballstemmer=2.0.0=pypi_0
sphinx=3.0.4=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=1.0.3=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.4=pypi_0
sqlite=3.31.1=h2a8f88b_1
tensorboard=1.15.0=pypi_0
tensorboard-plugin-wit=1.6.0.post3=pypi_0
tensorflow-estimator=1.15.1=pypi_0
tensorflow-gpu=1.15.0=pypi_0
tensorflow-gpu-estimator=2.2.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
tifffile=2020.5.30=pypi_0
times=0.7=pypi_0
tk=8.6.8=hfa6e2cd_0
toml=0.10.0=py37h28b3542_0
toolz=0.10.0=py_0
tornado=6.0.4=py37he774522_1
traitlets=4.3.3=pypi_0
typed-ast=1.4.1=py37he774522_0
urllib3=1.25.9=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_2
wcwidth=0.2.3=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py37_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h62dcd97_0
yaml=0.1.7=vc14h4cb57cf_1
zipp=3.1.0=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.4=ha9fde0e_3